refactor: extract block_production/rocksdb test modules into tests.rs
Move the large #[cfg(test)] modules out of mc-block-production lib.rs (2667->1115 LOC) and mc-db rocksdb/mod.rs (1185->1085 LOC) into sibling tests.rs files. Pure relocation: test code is unchanged, only de-indented by rustfmt. Verified with cargo check --tests on both crates.
